+++ api-gateway/src/main/java/ru/creditbank/apigateway/service/UserService.java
@@ -2,7 +2,6 @@ package ru.creditbank.apigateway.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import ru.creditbank.apigateway.dto.rq.LoginRqDto;