+++ api-gateway/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Add Swagger UI
 
 [0.0.5] - 2026-08-25
+### Added
+    - use FeignOpen Client as a proxy
+
 ### Changed
     - userService .findUserinfo() -> getInfo()
     - use HttpStatus instead HttpRequstResponse.SC_{CODE}